+++ noodles-sam/src/alignment/record/quality_scores.rs
@@ -1,3 +1,5 @@
+use std::io;
+
 /// Alignment record quality scores.
 pub trait QualityScores {
     /// Returns whether there are any scores.
@@ -7,11 +9,11 @@ pub trait QualityScores {
     fn len(&self) -> usize;
 
     /// Returns an iterator over scores.
-    fn iter(&self) -> Box<dyn Iterator<Item = u8> + '_>;
+    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<u8>> + '_>;
 }
 
 impl<'a> IntoIterator for &'a dyn QualityScores {
-    type Item = u8;
+    type Item = io::Result<u8>;
     type IntoIter = Box<dyn Iterator<Item = Self::Item> + 'a>;
 
     fn into_iter(self) -> Self::IntoIter {
@@ -28,7 +30,7 @@ impl QualityScores for Box<dyn QualityScores + '_> {
         (**self).len()
     }
 
-    fn iter(&self) -> Box<dyn Iterator<Item = u8> + '_> {
+    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<u8>> + '_> {
         (**self).iter()
     }
 }
@@ -38,7 +40,7 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_into_iter() {
+    fn test_into_iter() -> io::Result<()> {
         struct T(Vec<u8>);
 
         impl QualityScores for T {
@@ -50,16 +52,18 @@ mod tests {
                 self.0.len()
             }
 
-            fn iter(&self) -> Box<dyn Iterator<Item = u8> + '_> {
-                Box::new(self.0.iter().copied())
+            fn iter(&self) -> Box<dyn Iterator<Item = io::Result<u8>> + '_> {
+                Box::new(self.0.iter().copied().map(Ok))
             }
         }
 
         let quality_scores: &dyn QualityScores = &T(vec![45, 35, 43, 50]);
 
         assert_eq!(
-            quality_scores.into_iter().collect::<Vec<_>>(),
+            quality_scores.into_iter().collect::<io::Result<Vec<_>>>()?,
             [45, 35, 43, 50]
         );
+
+        Ok(())
     }
 }